cuda/hipblaslt: address remaining Copilot review findings
Five fixes from the 2026-08-11 review pass:

- vecdotq: pack the Q2_0 HIP-path qs bytes as unsigned. A byte >= 128
  shifted by 24 overflows the promoted int (UB); build qs0/qs1 as
  uint32_t so optimized HIP/MUSA builds cannot miscompile the high byte.

- common: fix fp32->E4M3 rounding at the subnormal/normal boundary. A
  subnormal mantissa that rounds to 8 is exactly 2^-6, the minimum
  normal (encoding 0x08); clamping it to mantissa 7 returned a
  non-nearest value.

- mmvq: make the opt-in activation-quant dedup cache stream-safe.
  Concurrent graph regions fork sibling matmuls onto separate streams;
  a hit could consume the cached q8_1 before the populating quantize
  (enqueued on the miss stream) completed. Record an event after the
  populating quantize and make hits on any other stream wait on it
  (same treatment the hipblaslt wcache got in 65cd5bd2). The event is
  destroyed with the context.

- hipblaslt: key both weight caches by (device, addr, N, K) instead of
  the raw device address, which multi-GPU runs and reshaped views can
  alias. Invalidation matches on address across devices (a false
  positive only costs a redundant re-requant).

- hipblaslt: give the Q1_0 route the Q2_0 route's VRAM preflights: the
  in-cache headroom check before hipMalloc and the op-level transient
  guard (weight-miss + activation + accumulator + workspace) so a
  near-full card falls back to mmq instead of OOM-aborting in the pool.

Verified on 1x R9700 (Bonsai-27B): test-backend-ops MUL_MAT passes
plain and with dedup+batch; PPL wiki.test 20 chunks identical with
dedup on/off (11.5603); int8/fp8 prefill routes run with the expected
uplift over the same-build no-route baseline (Q2_0 797->1178 t/s
pp1024 warm, Q1_0 1022->1169) and sane PPL (int8 11.7646, fp8 11.6337).

CCA (Claude Code Augmented)

+++ ggml/src/ggml-cuda/vecdotq.cuh
@@ -783,10 +783,11 @@ static __device__ __forceinline__ float vec_dot_q2_0_q8_1(
     // applying the -sum(u) offset once at the end via the q8_1 stored sum
     // (ds.y = d8 * sum(u)) -- no per-code subtract, plain shift/mask + dp4a.
     const int offset = iqs * 8;
-    const int qs0 = bq2_0->qs[offset + 0] | (bq2_0->qs[offset + 1] << 8) |
-                    (bq2_0->qs[offset + 2] << 16) | (bq2_0->qs[offset + 3] << 24);
-    const int qs1 = bq2_0->qs[offset + 4] | (bq2_0->qs[offset + 5] << 8) |
-                    (bq2_0->qs[offset + 6] << 16) | (bq2_0->qs[offset + 7] << 24);
+    // pack as unsigned: a byte >= 128 shifted by 24 overflows the promoted int (UB)
+    const uint32_t qs0 = (uint32_t) bq2_0->qs[offset + 0]        | ((uint32_t) bq2_0->qs[offset + 1] <<  8) |
+                        ((uint32_t) bq2_0->qs[offset + 2] << 16) | ((uint32_t) bq2_0->qs[offset + 3] << 24);
+    const uint32_t qs1 = (uint32_t) bq2_0->qs[offset + 4]        | ((uint32_t) bq2_0->qs[offset + 5] <<  8) |
+                        ((uint32_t) bq2_0->qs[offset + 6] << 16) | ((uint32_t) bq2_0->qs[offset + 7] << 24);
 
     int sumi = 0;   // = dot(c, u), c in {0,1,2,3}
 #pragma unroll
